pyproject.toml
[project]
name = "mp-fcb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.14",
    "greenlet>=3.2.3",
    "load-dotenv>=0.1.0",
    "pydantic>=2.11.7",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
